// app/requests/getForm/getForm.type.d.ts
import type { Inputs } from "../../shared/constructor/inputs";

// УКАЗЫВАЕМ ТИПЫ ИНПУТОВ
//     formData: Inputs;
// УКАЗЫВАЕМ ТИПЫ ИНПУТОВ

/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetFormInputsSchema {
  result: {
    /**
     * @minItems 1
     */
    formData: Inputs;
    step: number;
    type: "needRequired" | "allowedNewStep" | "addedNewFields";
    [k: string]: unknown;
  };
  status: "success";
  [k: string]: unknown;
}